Changes to support ephemeral volumes
Ephemeral volumes are tide to single application, and the data in that volume
will be deleted once the application exists. Ephemeral volume cannot be shared
by more that one application.

We control volume sharing by rejecting the multiple NodePublishVolume requests
coming for the same volume. And delete the volume on ControllerUnpublishVolume.